% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userInfo_add}
\alias{userInfo_add}
\title{写入用户便利店}
\usage{
userInfo_add(
  token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE",
  file_name = "data-raw/jh_dms_users.xlsx",
  app_id = "jhdms"
)
}
\arguments{
\item{token}{口令}

\item{file_name}{文件名}

\item{app_id}{程序名}
}
\value{
返回值
}
\description{
写入用户便利店
}
\examples{
userInfo_add()
}
